+++ mysql_class.py
@@ -1720,7 +1720,7 @@ class SlaveRep(Rep):
             self.secs_behind = int(data["Seconds_Behind_" + master])
 
         except (ValueError, TypeError):
-            self.secs_behind = data["Seconds_Behind_" + master]
+            self.secs_behind = "null" if self.secs_behind is None else "UNK"
 
     def start_slave(self):
 
@@ -1746,7 +1746,7 @@ class SlaveRep(Rep):
             self.secs_behind = int(data["Seconds_Behind_" + master])
 
         except (ValueError, TypeError):
-            self.secs_behind = data["Seconds_Behind_" + master]
+            self.secs_behind = "null" if self.secs_behind is None else "UNK"
 
     def show_slv_state(self):
 
@@ -1842,7 +1842,7 @@ class SlaveRep(Rep):
             self.secs_behind = int(data["Seconds_Behind_" + master])
 
         except (ValueError, TypeError):
-            self.secs_behind = data["Seconds_Behind_" + master]
+            self.secs_behind = "null" if self.secs_behind is None else "UNK"
 
         self.ssl_verify = data[master + "_SSL_Verify_Server_Cert"]
 
@@ -2044,7 +2044,6 @@ class SlaveRep(Rep):
 
         except (ValueError, TypeError):
             self.secs_behind = "null" if self.secs_behind is None else "UNK"
-#            self.secs_behind = data["Seconds_Behind_" + master]
 
     def get_time(self):